# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "lplsoc是go基础模板",
        "title": "lplsoc Example Api",
        "contact": {},
        "license": {
            "name": "Apache 2.01",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "v1.0"
    },
    "host": "localhost:8095",
    "basePath": "/api",
    "paths": {
        "/account/login": {
            "post": {
                "description": "login接口详情介绍",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "accountGroup"
                ],
                "summary": "login接口",
                "parameters": [
                    {
                        "description": "post传参",
                        "name": "dataInfo",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/loginMethod.SwagLogin"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/loginMethod.AuthError"
                        }
                    }
                }
            }
        },
        "/account/logout": {
            "delete": {
                "description": "用户退出接口",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "accountGroup"
                ],
                "summary": "用户退出登录",
                "parameters": [
                    {
                        "type": "string",
                        "description": "传token 例: Bearer eyJhbGzI1...ciOiJIU",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/loginMethod.LoginOut"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/loginMethod.AuthError"
                        }
                    }
                }
            }
        },
        "/account/testLogin": {
            "get": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "accountGroup"
                ],
                "summary": "用户登录验证",
                "parameters": [
                    {
                        "type": "string",
                        "description": "传token 例: Bearer eyJhbGzI1...ciOiJIU",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/loginMethod.LoginOut"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/loginMethod.AuthError"
                        }
                    }
                }
            }
        },
        "/account/userInfo": {
            "get": {
                "description": "详情",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "accountGroup"
                ],
                "summary": "获取用户信息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "传token 例: Bearer eyJhbGzI1...ciOiJIU",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/loginMethod.LoginOut"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/loginMethod.AuthError"
                        }
                    }
                }
            }
        },
        "/swag/aa/bb": {
            "get": {
                "description": "bb接口详情介绍",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "gavin"
                ],
                "summary": "bb接口",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/testExampleConf.Response"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/testExampleConf.HTTPError"
                        }
                    }
                }
            }
        },
        "/swag/aa/cc": {
            "get": {
                "description": "cc接口详情介绍",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "gavin"
                ],
                "summary": "cc接口",
                "responses": {
                    "200": {
                        "description": "answer--",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/testExampleConf.HTTPError"
                        }
                    },
                    "404": {
                        "description": "404报错"
                    },
                    "500": {
                        "description": "500报错"
                    }
                }
            }
        },
        "/swag/loginAA": {
            "post": {
                "description": "用于用户认证--接口详情介绍",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "bb"
                ],
                "summary": "登录接口 ---接口说明",
                "parameters": [
                    {
                        "type": "string",
                        "description": "学号",
                        "name": "stu_id",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "密码",
                        "name": "password",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/testExampleConf.HTTPSuccess"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/testExampleConf.HTTPError"
                        }
                    },
                    "500": {
                        "description": "未找到此用户"
                    }
                }
            }
        }
    },
    "definitions": {
        "loginMethod.AuthError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 1001
                },
                "message": {
                    "type": "string",
                    "example": "认证错误"
                }
            }
        },
        "loginMethod.LoginOut": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 200
                },
                "data": {
                    "type": "string",
                    "example": "success"
                },
                "msg": {
                    "type": "string",
                    "example": "null"
                }
            }
        },
        "loginMethod.SwagLogin": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 200
                },
                "data": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    },
                    "example": {
                        "Token": "eyJhbGciOiCI6IkpXVCJ9"
                    }
                },
                "msg": {
                    "type": "string",
                    "example": "null"
                }
            }
        },
        "testExampleConf.Data": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "integer",
                    "format": "int64",
                    "example": 1
                },
                "name": {
                    "type": "string",
                    "example": "gavin"
                },
                "title": {
                    "type": "string",
                    "example": "Object data"
                },
                "uuid": {
                    "type": "string",
                    "format": "uuid",
                    "example": "550e8400-e29b-41d4-a716-446655440000"
                }
            }
        },
        "testExampleConf.HTTPError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 400
                },
                "message": {
                    "type": "string",
                    "example": "status bad request"
                }
            }
        },
        "testExampleConf.HTTPSuccess": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 200
                },
                "data": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    },
                    "example": {
                        "Token": "eyJhbGciOiCI6IkpXVCJ9"
                    }
                },
                "msg": {
                    "type": "string",
                    "example": "null"
                }
            }
        },
        "testExampleConf.Response": {
            "type": "object",
            "properties": {
                "map_data": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    },
                    "example": {
                        "key": "value",
                        "key2": "value2"
                    }
                },
                "object": {
                    "$ref": "#/definitions/testExampleConf.Data"
                },
                "title": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "string"
                    },
                    "example": {
                        "en": "Map",
                        "kk": "Карталар",
                        "ru": "Карта"
                    }
                }
            }
        }
    }
}